feat(custom-audio-hook): create a custom hook for audio in useAudio.js, and refactor PokemonCry.js and ThemeSongs.js to implement the custom useAudio hook
