+++ pokemon-app/src/components/PokemonCry.js
@@ -1,28 +1,11 @@
-import React, { useState, useEffect, useRef } from "react";
-import { audioPlayAsync } from "../utils";
+import React, { useState, useEffect } from "react";
+import { useAudio } from "../hooks";
+// import { audioPlayAsync } from "../utils";
 
 const PokemonCry = ({ pokeSoundUrl, isPokeballRendering }) => {
   // Set the initial state for rendering the false.
   const [isPokemonsCries, setIsPokemonsCries] = useState(false);
-  // Initialize the audio reference.
-  const audioRef = useRef();
-
-  // This function is used to check if there are no errors on audio play.
-  const audioPlay = async (targetAudio) => {
-    try {
-      // Destructure the status and message.
-      const { status, message } = await audioPlayAsync(targetAudio);
-
-      // Check if the status had an error.
-      if (status === "error") {
-        // Send the error message for the catch block.
-        throw new Error(message);
-      }
-    } catch (error) {
-      // Display the error to the console.
-      console.error(error);
-    }
-  };
+  const [audioRef, audioPlay] = useAudio();
 
   useEffect(() => {
     // If sound url exists and the pokeball animation is not rendering,
